move tests to right places
